cursor関係をedit_editに移動
